requirements.txt
django==1.11
Pillow
django-imagekit
pilkit==2.0
django-taggit==0.22.2